Update hyeweon.cpp
week7

+++ 7-week7/1/hyeweon.cpp
@@ -10,7 +10,7 @@ int main()
 	for (int i = 0; i < n; i++)
 		cin >> blocks[i];
 
-	double sum = blocks[n - 1];
+	double sum = (double)blocks[n - 1];
 	int answer = 1;
 	for (int i = n - 2; i >= 0; i--) {
 		if ((blocks[i] - m > blocks[i + 1]) || (blocks[i] + m < blocks[i + 1])) {
